feat: just article nonactive can be dleted

+++ src/controllers/artikelController.js
@@ -126,14 +126,22 @@ class ArtikelController {
   }
 
   async getAllArtikel(req, res) {
-    const { page = 1, per_page = 15, keyword = "", status = "PUBLISH" } = req.query;
-  
+    const {
+      page = 1,
+      per_page = 15,
+      keyword = "",
+      status = "PUBLISH",
+    } = req.query;
+
     try {
       let response, totalArticles, lastPage, from, to;
-  
+
       if (page === "no") {
         // Ambil semua artikel tanpa pagination
-        response = await artikelRepository.getAllArtikelNoPagination(keyword, status);
+        response = await artikelRepository.getAllArtikelNoPagination(
+          keyword,
+          status
+        );
         totalArticles = response.length;
         lastPage = 1; // Tidak ada halaman terakhir untuk non-paginated
         from = 1;
@@ -141,14 +149,17 @@ class ArtikelController {
       } else {
         const currentPage = Number(page);
         const itemsPerPage = Number(per_page);
-  
+
         response = await artikelRepository.getAllArtikel(
           currentPage,
           itemsPerPage,
           keyword,
           status
         );
-        totalArticles = await artikelRepository.getTotalArtikel(keyword, status);
+        totalArticles = await artikelRepository.getTotalArtikel(
+          keyword,
+          status
+        );
         lastPage =
           totalArticles > 0 ? Math.ceil(totalArticles / itemsPerPage) : 1;
         from = (currentPage - 1) * itemsPerPage + 1;
@@ -157,7 +168,7 @@ class ArtikelController {
             ? Math.min(from + itemsPerPage - 1, totalArticles)
             : 0;
       }
-  
+
       res.status(200).json({
         status: 200,
         message: "Successfully retrieved all articles.",
@@ -219,6 +230,14 @@ class ArtikelController {
           status: 404,
           message: "Article not found. Cannot delete a non-existing article.",
         });
+
+      if (existArtikel.status !== "NONACTIVE") {
+        return res.status(403).json({
+          status: 403,
+          message:
+            "Article cannot be deleted because it is not in NONACTIVE status.",
+        });
+      }
 
       await artikelRepository.deleteArtikel(parseInt(id));
 
@@ -293,7 +312,9 @@ class ArtikelController {
       const files = req.files?.["media"] || [];
       let bannerId = null;
 
-      const existArtikel = await artikelRepository.findArtikelById(parseInt(id));
+      const existArtikel = await artikelRepository.findArtikelById(
+        parseInt(id)
+      );
       if (!existArtikel) {
         return res.status(404).json({
           status: 400,
